# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Attendee App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admins:
    get:
      operationId: listAdmins
      tags: [auth]
      summary: List all admin users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdmin
      tags: [auth]
      summary: Create a new admin user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminRequest"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [auth]
      summary: Delete an admin user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAdminResponse"
        "400":
          description: Cannot delete (self or last admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /settings/smtp:
    get:
      operationId: getSmtpSettings
      tags: [settings]
      summary: Get SMTP settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SMTP settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveSmtpSettings
      tags: [settings]
      summary: Save SMTP settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmtpSettingsRequest"
      responses:
        "200":
          description: Settings saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmtpSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/smtp/test:
    post:
      operationId: testSmtpSettings
      tags: [settings]
      summary: Test SMTP connection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/sms:
    get:
      operationId: getSmsSettings
      tags: [settings]
      summary: Get SMS provider settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SMS settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveSmsSettings
      tags: [settings]
      summary: Save SMS provider settings (KudiSMS token + sender ID)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsSettingsRequest"
      responses:
        "200":
          description: Settings saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsSettings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings/sms/test:
    post:
      operationId: testSmsSettings
      tags: [settings]
      summary: Verify KudiSMS token and fetch current balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Test result with balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsTestResponse"
        "400":
          description: Not configured or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - admin
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt
    AdminsResponse:
      type: object
      properties:
        admins:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
      required:
        - admins
    CreateAdminRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    DeleteAdminResponse:
      type: object
      properties:
        deleted:
          type: boolean
        admin:
          type: object
          properties:
            id:
              type: integer
            email:
              type: string
          required:
            - id
            - email
      required:
        - deleted
        - admin
...
    SmtpSettings:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        encryption:
          type: string
          enum: [tls, ssl, none]
        isConfigured:
          type: boolean
      required:
        - host
        - port
        - username
        - encryption
        - isConfigured
    SmtpSettingsRequest:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        username:
          type: string
          format: email
        password:
          type: string
        encryption:
          type: string
          enum: [tls, ssl, none]
      required:
        - host
        - port
        - username
        - password
        - encryption
...
    SmsSettings:
      type: object
      properties:
        provider:
          type: string
        senderId:
          type: string
        isConfigured:
          type: boolean
      required:
        - provider
        - senderId
        - isConfigured
    SmsSettingsRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        senderId:
          type: string
          minLength: 1
          maxLength: 11
      required:
        - token
        - senderId
    SmsTestResponse:
      type: object
      properties:
        message:
          type: string
        balance:
          type: number
      required:
        - message
        - balance
...