# lib/api-zod/src/generated/types/adminsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from "./adminUser.js";

export interface AdminsResponse {
  admins: AdminUser[];
}