# lib/api-zod/src/generated/types/deleteAdminResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */
import type { DeleteAdminResponseAdmin } from "./deleteAdminResponseAdmin.js";

export interface DeleteAdminResponse {
  deleted: boolean;
  admin: DeleteAdminResponseAdmin;
}