# lib/api-zod/src/generated/types/smsSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Event Attendee App API
 * OpenAPI spec version: 0.1.0
 */

export interface SmsSettings {
  provider: string;
  senderId: string;
  isConfigured: boolean;
}